# openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"Sentiment Analysis API","description":"An API for obtaining a sentiment analysis (percentage positive or negative) on a sentence that you provide.","version":"0.1-SNAPSHOT"},"servers":[{"url":"https://springboot-djl-demo-polyglot-demo.tap-next.blah.cloud","description":"Tanzu Application Platform"}],"tags":[{"name":"Springboot-Maven Deep Java Learning API Demo","description":"An API for obtaining a sentiment analysis (percentage positive or negative) on a sentence that you provide."}],"paths":{"/api/analyze":{"post":{"tags":["Sentence"],"summary":"Get a sentiment analysis on a sentence of text.","description":"The API returns a sentiment analysis based on a predefined ML model.","operationId":"analyze","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Sentence"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDescription"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SentimentAnalysis"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SentimentAnalysis"}}}},"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SentimentAnalysis"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SentimentAnalysis"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SentimentAnalysis"}}}}}}}},"components":{"schemas":{"Sentence":{"type":"object","properties":{"sentence":{"type":"string"}},"description":"The sentence to analyze"},"ErrorDescription":{"type":"object","properties":{"error":{"type":"string","description":"The error description, usually 'error'"},"status":{"type":"integer","description":"The HTTP Status code for this error.","format":"int32"}}},"SentimentAnalysis":{"type":"object","properties":{"sentence":{"type":"string","description":"The sentence to analyze"},"positive_probability":{"type":"string","description":"The probability that the sentence is positive in the format ##0.00000%"},"negative_probability":{"type":"string","description":"The probability that the sentence is negative in the format ##0.00000%"}},"description":"The result of a sentiment analysis"}}}}